network badge working on both platforms
